member2/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
web3>=6.19.0
pydantic>=2.7.1
pandas>=2.2.2
sse-starlette>=2.1.0
mcp>=1.0.0
